+++ common/dal-context/src/lib.rs
@@ -11,143 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use std::sync::Arc;
+mod metrics;
 
-use async_trait::async_trait;
-use common_dal2::error::Result as DalResult;
-use common_dal2::ops::OpDelete;
-use common_dal2::ops::OpRead;
-use common_dal2::ops::OpStat;
-use common_dal2::ops::OpWrite;
-use common_dal2::Accessor;
-use common_dal2::Layer;
-use common_dal2::Object;
-use common_dal2::Reader;
-use common_infallible::RwLock;
-
-#[derive(Clone, Debug, Default)]
-pub struct DalMetrics {
-    /// Read bytes.
-    pub read_bytes: usize,
-    /// Seek times of read.
-    pub read_seeks: usize,
-    /// Cost(in ms) of read bytes.
-    pub read_byte_cost_ms: usize,
-    /// Cost(in ms) of seek by reading.
-    pub read_seek_cost_ms: usize,
-    /// Bytes written by data access layer
-    pub write_bytes: usize,
-    /// Number of rows written
-    pub write_rows: usize,
-    /// Number of partitions scanned, after pruning
-    pub partitions_scanned: usize,
-    /// Number of partitions, before pruning
-    pub partitions_total: usize,
-}
-
-#[derive(Clone)]
-pub struct DalContext {
-    inner: Arc<dyn Accessor>,
-    metrics: Arc<RwLock<DalMetrics>>,
-}
-
-impl DalContext {
-    pub fn new(inner: Arc<dyn Accessor>) -> Self {
-        DalContext {
-            inner,
-            metrics: Arc::new(Default::default()),
-        }
-    }
-
-    /// Increment read bytes.
-    pub fn inc_read_bytes(&self, bytes: usize) {
-        if bytes > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.read_bytes += bytes;
-        }
-    }
-
-    /// Increment write bytes.
-    pub fn inc_write_bytes(&self, bytes: usize) {
-        if bytes > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.write_bytes += bytes;
-        }
-    }
-
-    /// Increment read seek times.
-    pub fn inc_read_seeks(&self) {
-        let mut metrics = self.metrics.write();
-        metrics.read_seeks += 1;
-    }
-
-    /// Increment cost for reading bytes.
-    pub fn inc_read_byte_cost_ms(&self, cost: usize) {
-        if cost > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.read_byte_cost_ms += cost;
-        }
-    }
-
-    //// Increment cost for reading seek.
-    pub fn inc_read_seek_cost_ms(&self, cost: usize) {
-        if cost > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.read_seek_cost_ms += cost;
-        }
-    }
-
-    //// Increment numbers of rows written
-    pub fn inc_write_rows(&self, rows: usize) {
-        if rows > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.write_rows += rows;
-        }
-    }
-
-    //// Increment numbers of partitions scanned
-    pub fn inc_partitions_scanned(&self, partitions: usize) {
-        if partitions > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.partitions_scanned += partitions;
-        }
-    }
-
-    //// Increment numbers of partitions (before pruning)
-    pub fn inc_partitions_total(&self, partitions: usize) {
-        if partitions > 0 {
-            let mut metrics = self.metrics.write();
-            metrics.partitions_total += partitions;
-        }
-    }
-
-    pub fn get_metrics(&self) -> DalMetrics {
-        self.metrics.read().clone()
-    }
-}
-
-impl Layer for DalContext {
-    fn layer(&self, inner: Arc<dyn Accessor>) -> Arc<dyn Accessor> {
-        Arc::new(DalContext::new(inner))
-    }
-}
-
-#[async_trait]
-impl Accessor for DalContext {
-    async fn read(&self, args: &OpRead) -> DalResult<Reader> {
-        // TODO(xuanwo): Implement context callback reader to collect metrics.
-        self.inner.read(args).await
-    }
-    async fn write(&self, r: Reader, args: &OpWrite) -> DalResult<usize> {
-        self.inner.write(r, args).await.map(|n| {
-            self.inc_write_bytes(n);
-            n
-        })
-    }
-    async fn stat(&self, args: &OpStat) -> DalResult<Object> {
-        self.inner.stat(args).await
-    }
-    async fn delete(&self, args: &OpDelete) -> DalResult<()> {
-        self.inner.delete(args).await
-    }
-}
+mod context;
+pub use context::DalContext;
